Fix parsing of bed and aed files

+++ gens/commands/load.py
@@ -10,23 +10,24 @@ from flask import current_app as app
 from flask.cli import with_appcontext
 from pymongo.database import Database
 
+from gens.config import settings
 from gens.db import (
     ANNOTATIONS_COLLECTION,
     CHROMSIZES_COLLECTION,
     SAMPLES_COLLECTION,
     TRANSCRIPTS_COLLECTION,
     create_index,
+    get_db_connection,
     get_indexes,
     register_data_update,
     store_sample,
 )
 from gens.load import (
-    ParserError,
     build_chromosomes_obj,
     build_transcripts,
     get_assembly_info,
     parse_annotation_entry,
-    parse_annotation_file,
+    read_annotation_file,
     update_height_order,
 )
 from gens.models.annotation import AnnotationRecord
@@ -148,10 +149,16 @@ def sample(
     required=True,
     help="Genome build",
 )
-@with_appcontext
-def annotations(file: str, genome_build: GenomeBuild):
+@click.option(
+    "-h",
+    "--header",
+    "has_header",
+    is_flag=True,
+    help="If bed file contains a header",
+)
+def annotations(file: str, genome_build: GenomeBuild, has_header: bool):
     """Load annotations from file into the database."""
-    db = app.config["GENS_DB"]
+    db = get_db_connection(settings.gens_db, db_name=settings.gens_dbname)
     # if collection is not indexed, create index
     if len(get_indexes(db, ANNOTATIONS_COLLECTION)) == 0:
         create_index(db, ANNOTATIONS_COLLECTION)
@@ -166,16 +173,16 @@ def annotations(file: str, genome_build: GenomeBuild):
         LOG.info("Processing %s", annot_file)
         # base the annotation name on the filename
         annotation_name = annot_file.name[: -len(annot_file.suffix)]
-        parser = parse_annotation_file(annot_file, file_format=annot_file.suffix[1:])
         parsed_annotations: list[AnnotationRecord] = []
-        for entry in parser:
-            try:
-                entry_obj = parse_annotation_entry(entry, genome_build, annotation_name)
+        for entry in read_annotation_file(
+            annot_file, annot_file.suffix[1:], has_header
+        ):
+            entry_obj = parse_annotation_entry(entry, genome_build, annotation_name)
+            if entry_obj is not None:
                 parsed_annotations.append(entry_obj)
-            except ParserError as err:
-                LOG.warning(str(err))
-                continue
 
+        if len(parsed_annotations) == 0:
+            raise ValueError("Something went wrong parsing the annotaions file.")
         # Remove existing annotations in database
         LOG.info("Remove old entry in the database")
         db[ANNOTATIONS_COLLECTION].delete_many({"source": annotation_name})
@@ -185,7 +192,9 @@ def annotations(file: str, genome_build: GenomeBuild):
         LOG.info("Update height order")
         # update the height order of annotations in the database
         update_height_order(db, annotation_name)
-        register_data_update(ANNOTATIONS_COLLECTION, name=annotation_name)
+        register_data_update(
+            db=db, track_type=ANNOTATIONS_COLLECTION, name=annotation_name
+        )
     click.secho("Finished loading annotations ✔", fg="green")
 
 
@@ -226,7 +235,7 @@ def transcripts(file: str, mane: str, genome_build: GenomeBuild):
 
     LOG.info("Add transcripts to database")
     db[TRANSCRIPTS_COLLECTION].insert_many(transcripts_obj)
-    register_data_update(TRANSCRIPTS_COLLECTION)
+    register_data_update(db, TRANSCRIPTS_COLLECTION)
     click.secho("Finished loading transcripts ✔", fg="green")
 
 
@@ -271,10 +280,12 @@ def chromosome_info(genome_build: GenomeBuild, timeout: int):
 
     # remove old entries
     res = db[CHROMSIZES_COLLECTION].delete_many({"genome_build": int(genome_build)})
-    LOG.info("Removed %d old entries with genome build: %s", res.deleted_count, genome_build)
+    LOG.info(
+        "Removed %d old entries with genome build: %s", res.deleted_count, genome_build
+    )
     # insert collection
     LOG.info("Add chromosome info to database")
     db[CHROMSIZES_COLLECTION].insert_many([chr.model_dump() for chr in chromosomes_data])
-    register_data_update(CHROMSIZES_COLLECTION)
+    register_data_update(db, CHROMSIZES_COLLECTION)
     # build cytogenetic data
     click.secho("Finished updating chromosome sizes ✔", fg="green")
